requirements.txt
pandas
numpy
seaborn
google-cloud-bigquery
db-dtypes
google-cloud-bigquery-storage
-e .